Code style, part 2/2: code quality, AOT, interpolation, and enforcement (#1343)
* Code quality: remove dead usings/casts, add modifiers, readonly fields

Auto-applied with 'dotnet format' and verified (build green, tests pass):
- IDE0005: remove unnecessary using directives
- IDE0004: remove unnecessary casts (all identity/implicit conversions)
- IDE0040: add explicit accessibility modifiers matching the existing default
- IDE0044: mark fields readonly where never reassigned
- IDE1005: use conditional delegate invocation (?.Invoke) for events

IDE0059 (unnecessary assignment) was deliberately excluded: its fixer
rewrites named, commented reads into bare '_ =' discards and misfired on
interop structs, so it stays at suggestion rather than being auto-applied.


* Fix AOT/trim warnings in PropVariant and MfActivate blob marshalling

Annotate the generic parameter of both GetBlobAsArrayOf<T> helpers with
[DynamicallyAccessedMembers] so the constructors the trimmer must preserve
are declared, clearing IL2087 (Activator.CreateInstance) and IL2091
(Marshal.PtrToStructure<T>). Keeps NAudio.Wasapi clean under the AOT/trim
analyzers that back the AOT smoke test.


* Use string interpolation instead of String.Format in library code

Convert 46 String.Format/string.Format calls across the NAudio libraries to
interpolated strings for readability. Behaviour-preserving:
- The one call passing CultureInfo.InvariantCulture is left as String.Format
  so its explicit culture is not silently changed to the current culture.
- Commented-out String.Format references are left untouched.

Build green, all NAudio.Core.Tests pass.


* Enforce curated code-style rules at build time

Promote the now-conformant code-style rules from suggestion to warning so
TreatWarningsAsErrors steers contributors automatically:
- .globalconfig: IDE0055 (formatting), IDE0161 (file-scoped namespaces),
  IDE0004 (casts), IDE0040 (accessibility), IDE0044 (readonly),
  IDE1005 (?.Invoke), IDE0090 (target-typed new), IDE0063 (using
  declarations), IDE0031 (null propagation)
- .editorconfig: IDE0005 (unnecessary usings) scoped to src/, since it
  requires GenerateDocumentationFile=true which only the libraries set
- Remove a few usings left unnecessary by the interpolation sweep

Adds .git-blame-ignore-revs listing the two mechanical reformat commits, and
records the sweep in RELEASE_NOTES.


* Restore non-ASCII characters mangled by the reformat pass

Four files were Windows-1252 encoded (no BOM); 'dotnet format' read them as
UTF-8, replaced the undecodable bytes with U+FFFD, and re-saved as UTF-8.
Restore the intended characters, now as proper UTF-8:
- 'Copyright (c) 2007 Mark Heath' in the three AboutForm.cs sample files
- the 'far field' curly quotes in AudioStreamCategory.cs


* Convert remaining Windows-1252 source files to UTF-8

Defensive follow-up: re-encode the last non-UTF-8 text files in the repo to
UTF-8 (with BOM) so a future 'dotnet format' run can't mangle them the way the
AboutForm files were. All contained a Windows-1252 copyright symbol:
- AboutForm.Designer.cs in the three sample apps (the designer halves that the
  reformat pass happened to skip)
- InstallScript.nsi in AudioFileInspector and MidiFileConverter (UTF-8 BOM is
  also how NSIS 3 detects UTF-8)

No remaining non-UTF-8 text files in the tree. Build green.


* Remove obsolete NSIS installer scripts

The InstallScript.nsi files for AudioFileInspector and MidiFileConverter date
from ~2007-2009 and no longer work for the current samples: they gate on a
.NET Framework 2.0 CLR check (the apps now target net9.0-windows), link to a
dead MSDN URL, and bundle a single monolithic NAudio.dll rather than the
split NAudio 3 assemblies plus the modern .NET host/deps files. They're not
referenced by any project, solution, or CI build. NAudio ships via NuGet, so
maintaining sample-app installers adds no value.


* Fix CI build: pin label indentation and a Release-only unnecessary using

Two formatting errors only surfaced under the CI build (Release, and a newer
analyzer than the local SDK), both now reproduced and fixed:

- IDE0055 on the WdlResampler 'again:' goto label: SDK versions disagree on the
  default label-indentation style. Pin csharp_indent_labels = no_change so the
  single ported label is left as written rather than flagged.
- IDE0005 unnecessary using in AudioFileFormatFinder: System.Runtime.Interop-
  Services was only referenced from a #if DEBUG catch, so it was unused in
  Release. Drop the using and fully-qualify the one ExternalException reference.

Verified with full-solution builds in BOTH Debug and Release; tests pass.


* Point .git-blame-ignore-revs at the squashed mechanical commit

Part 1/2 (#1344) was squash-merged into a single commit, so replace the two
old per-commit SHAs with that squash SHA.


---------

+++ src/NAudio.Core/Codecs/ALawEncoder.cs
@@ -46,7 +46,7 @@ public static class ALawEncoder
             sample = cClip;
         if (sample >= 256)
         {
-            exponent = (int)ALawCompressTable[(sample >> 8) & 0x7F];
+            exponent = ALawCompressTable[(sample >> 8) & 0x7F];
             mantissa = (sample >> (exponent + 3)) & 0x0F;
             compressedByte = (byte)((exponent << 4) | mantissa);
         }
